Se actualiza controlador usuario

+++ ESFE AGAPE BODEGA.DTOs/UsuarioDTOs/SearchQueryUsuarioDTO.cs
@@ -21,6 +21,6 @@ namespace ESFE_AGAPE_BODEGA.DTOs.UsuarioDTOs
 		/// 1 = No se cuenta los resultados de la busqueda
 		/// 2= Cuenta los resultados de la busqueda
 		/// </summary>
-		public byte SendRowCount { get; set; }
+		public int SendRowCount { get; set; }
 	}
 }
